+++ src/phishbench/utils/phishbench_globals.py
@@ -4,6 +4,9 @@ This module contains the global configuration for PhishBench
 import argparse
 import configparser
 import logging
+import os
+from io import TextIOBase
+from typing import Optional
 
 # pylint: disable=global-statement
 # pylint: disable=invalid-name
@@ -12,21 +15,25 @@ import logging
 args = None
 config = configparser.ConfigParser()
 logger: logging.Logger = logging.getLogger('root')
-summary = None
+summary: Optional[TextIOBase] = None
+output_dir = ""
 
 
-def setup_parser():
+def parse_args():
     """
     Sets up the argument parser
     """
     global args
+    global output_dir
     parser = argparse.ArgumentParser(description='PhishBench Basic Experiment Script')
     parser.add_argument("--version", help="Display the PhishBench version number and exit", action="store_true")
     parser.add_argument("-f", "--config_file", help="The config file to use", type=str, default='Default_Config.ini')
     parser.add_argument("-v", "--verbose", help="Increase output verbosity", action="store_true")
-    parser.add_argument("-o", "--output_input_dir", help="Output/input directory", type=str, default="Data_Dump")
+    parser.add_argument("-o", "--output_input_dir", help="Output/input directory",
+                        type=str, default="PhishBench Output")
     parser.add_argument("-c", "--ignore_confirmation", help="Do not wait for user's confirmation", action="store_true")
     args = parser.parse_args()
+    output_dir = args.output_input_dir
 
 
 def setup_logger(filename='phishbench.log'):
@@ -50,7 +57,8 @@ def setup_logger(filename='phishbench.log'):
         console_handler.setFormatter(formatter)
         logger.addHandler(console_handler)
 
-    file_handler = logging.FileHandler(filename)
+    log_path = os.path.join(output_dir, filename)
+    file_handler = logging.FileHandler(log_path)
     file_handler.setFormatter(formatter)
     logger.addHandler(file_handler)
 
@@ -72,8 +80,12 @@ def initialize(config_file):
     global config
     global summary
 
+    if not os.path.isdir(output_dir):
+        os.mkdir(output_dir)
+
     config.read(config_file)
-    summary = open(config["Summary"]["Path"], 'w')
+    summary_path = os.path.join(output_dir, config["Summary"]["Path"])
+    summary = open(summary_path, 'w')
     setup_logger()